requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
babelfish==0.6.1
beautifulsoup4==4.14.3
certifi==2026.2.25
chardet==7.4.3
charset-normalizer==3.4.7
click==8.3.2
click-option-group==0.5.9
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dogpile.cache==1.5.0
enzyme==0.5.2
frozenlist==1.8.0
greenlet==3.4.0
groq==0.37.1
guessit==3.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.1.1
knowit==0.5.11
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.147
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==1.26.4
orjson==3.11.8
packaging==24.2
platformdirs==4.9.6
PlexAPI==4.18.1
propcache==0.4.1
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
pymediainfo==7.0.1
pysubs2==1.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
rebulk==3.2.0
requests==2.33.1
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
srt==3.5.3
stevedore==5.7.0
subliminal==2.6.0
tenacity==8.5.0
tomlkit==0.14.0
trakit==0.2.5
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
